requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
awsebcli==3.20.10
bert-score==0.3.13
blessed==1.20.0
blinker==1.4
botocore==1.31.85
cement==2.8.2
cer==1.2.0
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cycler==0.12.1
datasets==2.19.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.13
djangorestframework==3.15.1
docutils==0.17.1
entmax==1.3
evaluate==0.4.2
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
gyp==0.1
httplib2==0.20.2
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipynb==0.5.1
ipython==8.18.1
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonargparse==3.13.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Levenshtein==0.25.1
lightning-utilities==0.11.2
lxml==5.2.2
Markdown==3.3.6
MarkupSafe==2.1.5
matplotlib==3.4.3
matplotlib-inline==0.1.7
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.10.1
pexpect==4.9.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
plotly==5.21.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.2.4
pytz==2024.1
PyYAML==5.4.1
pyzmq==26.0.3
rapidfuzz==3.9.0
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
roman==3.3
rpds-py==0.18.1
rsa==4.8
s3transfer==0.5.0
sacrebleu==2.4.2
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.1
semantic-version==2.8.5
sentencepiece==0.1.99
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.3.0
termcolor==1.1.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchmetrics==0.10.3
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
triton==2.3.0
ttystatus==0.38
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unbabel-comet==2.2.2
urllib3==1.26.18
virtualenv==20.26.2
wadllib==1.3.6
wcwidth==0.1.9
xxhash==3.4.1
yarl==1.9.4
zipp==1.0.0
